requirements.txt
notebooklm-py>=0.3.3
httpx>=0.27.0
